Updates to PHPUnit 4
Updates PHPUnit dependency to ~4.0. Removes dependency for
phpunit-selenium and phpunit-story as they appeared to not be used
anywhere.

Customizes bin location of compsoer to ./bin to make it easier to
run composer's PHPUnit. Adds bin to the .gitignore. Updates travis
config to install composer's dev dependencies and to run  composer's
PHPUnit.